+++ request_api_helper/lib/helper.dart
@@ -1,0 +1,36 @@
+void timeTracker(name, Function function) async {
+  final now = DateTime.now();
+  await function();
+  print('$name ' + DateTime.now().difference(now).inMilliseconds.toString() + ' Millisecond');
+}
+
+notNullFill(dynamic base, dynamic newData) {
+  if (newData != null) {
+    return newData;
+  }
+  return base;
+}
+
+handlingData(data, {bool debug = false}) {
+  if (data is Map) {
+    if (debug) {
+      print(data.toString() + ' // Map');
+    }
+  } else if (data is List) {
+    if (debug) {
+      print(data.toString() + ' // List');
+    }
+  } else if (data is String) {
+    if (debug) {
+      print(data + ' // String');
+    }
+  } else if (data is int) {
+    if (debug) {
+      print(data.toString() + ' // integer');
+    }
+  } else if (data is double) {
+    if (debug) {
+      print(data.toString() + ' // double');
+    }
+  }
+}