+++ request_api_helper/lib/model/server_switcher_model.dart
@@ -1,7 +1,0 @@
-class ServerSwitcherData {
-  String? nameServer;
-
-  /// base Domain Example : https://mydomain.com
-  String? linkServer;
-  ServerSwitcherData({this.linkServer, this.nameServer});
-}